+++ sentry/src/main/java/io/sentry/SentryEvent.java
@@ -1,6 +1,8 @@
 package io.sentry;
 
 import io.sentry.protocol.*;
+import io.sentry.util.CollectionUtils;
+import java.util.ArrayList;
 import java.util.Date;
 import java.util.HashMap;
 import java.util.List;
@@ -167,7 +169,7 @@ public final class SentryEvent extends SentryBaseEvent implements IUnknownProper
   }
 
   public void setFingerprints(final @Nullable List<String> fingerprint) {
-    this.fingerprint = fingerprint;
+    this.fingerprint = fingerprint != null ? new ArrayList<>(fingerprint) : null;
   }
 
   @ApiStatus.Internal
@@ -187,7 +189,7 @@ public final class SentryEvent extends SentryBaseEvent implements IUnknownProper
   }
 
   public void setModules(final @Nullable Map<String, String> modules) {
-    this.modules = modules;
+    this.modules = CollectionUtils.newHashMap(modules);
   }
 
   public void setModule(final @NotNull String key, final @NotNull String value) {