+++ sentry/src/main/java/io/sentry/protocol/User.java
@@ -125,11 +125,7 @@ public final class User implements Cloneable, IUnknownPropertiesConsumer {
    * @param other the other user related data..
    */
   public void setOthers(final @Nullable Map<String, @NotNull String> other) {
-    if (other != null) {
-      this.other = new ConcurrentHashMap<>(other);
-    } else {
-      this.other = null;
-    }
+    this.other = CollectionUtils.newConcurrentHashMap(other);
   }
 
   /**
@@ -164,8 +160,8 @@ public final class User implements Cloneable, IUnknownPropertiesConsumer {
   public @NotNull User clone() throws CloneNotSupportedException {
     final User clone = (User) super.clone();
 
-    clone.other = CollectionUtils.shallowCopy(other);
-    clone.unknown = CollectionUtils.shallowCopy(unknown);
+    clone.other = CollectionUtils.newConcurrentHashMap(other);
+    clone.unknown = CollectionUtils.newConcurrentHashMap(unknown);
 
     return clone;
   }