+++ sentry/src/test/java/io/sentry/protocol/RequestTest.kt
@@ -1,5 +1,6 @@
 package io.sentry.protocol
 
+import java.util.Collections
 import kotlin.test.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertNotNull
@@ -64,6 +65,39 @@ class RequestTest {
         request.others = null
 
         assertNull(request.others)
+    }
+
+    @Test
+    fun `when setEnvs receives immutable map as an argument, its still possible to add more env to the request`() {
+        val request = Request().apply {
+            envs = Collections.unmodifiableMap(mapOf("env1" to "value1"))
+            envs!!["env2"] = "value2"
+        }
+        assertNotNull(request.envs) {
+            assertEquals(mapOf("env1" to "value1", "env2" to "value2"), it)
+        }
+    }
+
+    @Test
+    fun `when setOther receives immutable map as an argument, its still possible to add more others to the request`() {
+        val request = Request().apply {
+            others = Collections.unmodifiableMap(mapOf("key1" to "value1"))
+            others!!["key2"] = "value2"
+        }
+        assertNotNull(request.others) {
+            assertEquals(mapOf("key1" to "value1", "key2" to "value2"), it)
+        }
+    }
+
+    @Test
+    fun `when setHeaders receives immutable map as an argument, its still possible to add more headers to the request`() {
+        val request = Request().apply {
+            headers = Collections.unmodifiableMap(mapOf("key1" to "value1"))
+            headers!!["key2"] = "value2"
+        }
+        assertNotNull(request.headers) {
+            assertEquals(mapOf("key1" to "value1", "key2" to "value2"), it)
+        }
     }
 
     private fun createRequest(): Request {